// cosmowcl/src/app/shared/models/api/cosmopratiche/praticaInRelazione.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * Pratiche
 * API per la gestione pratiche 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Pratica } from './pratica.d';
import { TipoRelazionePraticaPratica } from './tipoRelazionePraticaPratica.d';


export interface PraticaInRelazione { 
    pratica: Pratica;
    tipoRelazione: TipoRelazionePraticaPratica;
}